+++ tests/test_coro_mock.py
@@ -19,3 +19,21 @@ class CoroMockTest(unittest.TestCase):
 
         cm.assert_called_once_with(3, 4)
         self.assertEqual('123', result)
+
+    def test_setting_side_effect(self):
+        from tests.mock_responses import CoroMock
+
+        cm = CoroMock()
+        cm.coro.side_effect = ['123', '456', IOError('oops')]
+
+        self.assertEqual('123', self.loop.run_until_complete(cm(3, 4)))
+        self.assertEqual('456', self.loop.run_until_complete(cm(3, 4)))
+
+        with self.assertRaises(IOError):
+            self.loop.run_until_complete(cm(3, 4))
+
+        # A generator is not allowed to raise StopIteration by itself,
+        # so the StopIteration caused by side_effect being exhausted
+        # results in a RuntimeError.
+        with self.assertRaises(RuntimeError):
+            self.loop.run_until_complete(cm(3, 4))